requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
langchain==0.2.1
langchain-community==0.2.1
langchain-text-splitters==0.2.0
faiss-cpu==1.7.2
huggingface-hub==0.23.1
transformers==4.18.0
PyMuPDF==1.24.4
python-docx==0.8.11
sentence-transformers==2.2.2
gradio==3.1.4
httpx==0.23.0
httpcore==0.15.0